twitter-trending/data-generator/requirements.txt
kafka-python==2.0.2
requests==2.27.1
schedule==1.1.0
tweepy==4.5.0
python-dotenv==0.19.2